backend/requirements/base.txt
Django==3.2.16
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.6
django-cors-headers==4.0.0
django-filter==22.1
Pillow==9.5.0
redis==4.5.4
django-redis==5.2.0
boto3==1.24.59
django-storages==1.13.1
celery==5.2.7
python-dotenv==1.0.0
requests==2.28.2
drf-spectacular==0.25.1 